RS_Checkpoint_2/account-service/requirements.txt
fastapi
uvicorn
python-jose
requests
python-dotenv
